==> 02_fundamentos_pandas/notebook/02_base_de_dados.ipynb <==
# Fundamentos Pandas

import pandas as pd

pd.read_csv('../dados/aluguel.csv', sep = ';')

dados = pd.read_csv('../dados/aluguel.csv', sep = ';')
dados

type(dados)

# mostra informações sobre o objeto
dados.info()

## Informações Gerais Sobre a Base de Dados

dados.dtypes

pd.DataFrame(dados.dtypes, columns=['Tipos de Dados'])

tipos_de_dados = pd.DataFrame(dados.dtypes, columns=['Tipos de Dados'])
tipos_de_dados.columns.name = 'Variáveis'
tipos_de_dados

print(f'A Base de Dados apresenta {dados.shape[0]} imóveis e {dados.shape[1]} variáveis')

## Criando um Data Frame

import pandas as pd
data = [['Fulano', 12, 7.0, True],
        ['Sicrano', 15, 3.5, False], 
        ['Beltrano', 18, 9.3, True]]
dados = pd.DataFrame(data, 
        columns = ['Aluno', 'Idade', 'Nota', 'Aprovado'])
dados
