==> 02_fundamentos_pandas/notebook/03_importando_dados.ipynb <==
# Importando Diferentes Dados

import pandas as pd

## Lendo um arquivo json

json = open('../dados/aluguel.json')
print(json.read())

## Lendo um arquivo json com o pandas

pd.read_json('../dados/aluguel.json')
# mostra as 10 primeiras observações
# o pandas organiza as colunas de forma alfabética

# pode-se colocar dentro de uma variável
df_json = pd.read_json('../dados/aluguel.json')

## Lendo um arquivo txt

txt = open('../dados/aluguel.txt')
print(txt.read())

## Lendo um arquivo txt com o pandas

df_txt = pd.read_table('../dados/aluguel.txt')
df_txt

## Lendo um arquivo do excel xlsx com o pandas

# arquivos xlsx não possuem uma leitura nativa como os outros tipos
df_xlsx = pd.read_excel('../dados/aluguel.xlsx')
df_xlsx

## Lendo um Data Frame em um endereço html

df_html = pd.read_html('https://pt.wikipedia.org/wiki/Lista_das_maiores_empresas_do_Brasil_por_faturamento')

len(df_html) # mostra o nº de df presente no endereço

df_html[1]

df_html[2]
