==> 02_fundamentos_pandas/notebook/04_selecionando_uma_variavel.ipynb <==
# Relatório de Análise II

## Tipos de Imóveis

import pandas as pd

### Lendo dados csv trocando separador e lendo os 10 primeiros

dados = pd.read_csv('../dados/aluguel.csv', sep=';')

dados.head(10)

### Selecionando apenas a variável 'Tipo'

# dados['Tipo'] ou 
dados.Tipo # essa conotação só serve para um nome de variável válido

tipo_de_imovel = dados['Tipo']

type(tipo_de_imovel)

* existem alguns métodos para trabalhar com uma variável
```
df.value_counts() # mostra as variáveis disponíveis e qtas vezes elas aparecem
df.unique() # mostra as variáveis disponíveis
```

tipo_de_imovel.drop_duplicates() # mostra apenas a 1ª ocorrência e elimina as outras

### Fixando o método na variável

tipo_de_imovel.drop_duplicates(inplace=True)
tipo_de_imovel

## Organizando a Visualização

tipo_de_imovel = pd.DataFrame(tipo_de_imovel)
tipo_de_imovel
# aqui é mostrado onde encontrou os índices pela primeira vez

### Modificando o index
* iremos portanto reorganizar o index para fazer mais sentido

tipo_de_imovel.index
# observa os valores do index

tipo_de_imovel.shape[0]
# mostra o nº de linhas do data frame

range(tipo_de_imovel.shape[0])
# vendo o range do index

# usando o for para visualizar o range
for i in range(tipo_de_imovel.shape[0]):
    print(i)

# tipo_de_imovel.index = range(1, 23)
# sabendo o range, é possível numerar manualmente
# mas numera-se automaticamente para facilitar manutenção futura
tipo_de_imovel.index = range(tipo_de_imovel.shape[0])
tipo_de_imovel

tipo_de_imovel.index
# aqui pode-se ver onde começa, onde termina e o passo

# renomeando coluna index
tipo_de_imovel.columns.name = 'Id'

### Resultado final

tipo_de_imovel
