==> 02_fundamentos_pandas/notebook/05_trabalhando_com_series.ipynb <==
import pandas as pd

# Series

data = [1, 2, 3, 4, 5]

s = pd.Series(data) # cria uma serie
s

## Criando um Rótulo para o Index

index = ['Linha' + str(i) for i in range(5)]
index

## Passando esse Rótulo para a Serie

s = pd.Series(data=data, index=index)
s

## Passando um Dicionário para a Serie

data = {'Linha' + str(i): i + 1 for i in range(5)}
data

# nesse caso não é necessário passar o index,
# pois ele pega do dicionário
s = pd.Series(data)
s

## Operações com Series

### Uma Serie com uma Constante

s1 = s + 2
s1

### Operação entre Series

s2 = s + s1
s2
