==> 02_fundamentos_pandas/notebook/06_concatenando_dataframes.ipynb <==
import pandas as pd

# Concatenar DataFrames

## Criando um DataFrame a partir de uma lista

data = [[1, 2, 3], 
        [4, 5, 6], 
        [7, 8, 9]]
data

df1 = pd.DataFrame(data)
df1

### Renomeando o Index

index = ['Linha' + str(i) for i in range(3)]
index

df1 = pd.DataFrame(data=data, index=index)
df1

### Renomeando as Colunas

columns = ['Coluna' + str(i) for i in range(3)]
columns

df1 = pd.DataFrame(data=data, index=index, columns=columns)
df1

## Criando um DataFrame de um Dicionário

data = {
    'Coluna0': {'Linha0': 1, 'Linha1': 4, 'Linha2': 7},
    'Coluna1': {'Linha0': 2, 'Linha1': 5, 'Linha2': 8},
    'Coluna2': {'Linha0': 3, 'Linha1': 6, 'Linha2': 9}
}
data

df2 = pd.DataFrame(data)
df2

## Criando um DataFrame de Tuplas

data = [(1, 2, 3), 
        (4, 5, 6), 
        (7, 8, 9)]
data

df3 = pd.DataFrame(data=data, index=index, columns=columns)
df3

## Concatenando DataFrames

df1[df1 > 0] = 'A'
df1

df2[df2 > 0] = 'B'
df2

df3[df3 > 0] = 'C'
df3

### Empilhando

df4 = pd.concat([df1, df2, df3])
df4
# empilha um dataframe em cima do outro
# usa como variável de link as colunas
# permite linhas com mesmos rótulos

### Lado a Lado

df4 = pd.concat([df1, df2, df3], axis=1)
df4
# para concatenação lado a lado, muda-se o parâmetro axis para 1
# o padrão é axis=0, para empilhar
# nesse caso permite-se colunas com mesmos rótulos
