==> 02_fundamentos_pandas/notebook/08_exportando_dados.ipynb <==
# Relatório de Análise III

## Imóveis Residenciais

import pandas as pd

dados = pd.read_csv('../dados/aluguel.csv', sep=';')
dados.head(10)

## Filtrando Imóveis pela Coluna Tipo

* queremos apenas imóveis do tipo residencial
* para isso, podemos usar alguns tipos de seleção na coluna Tipo

### Usando unique()

dados.Tipo.unique()

### Usando value_counts()

dados.Tipo.value_counts()

### Usando drop_duplicates()

dados.Tipo.drop_duplicates()

### Colocando drop_duplicates() em uma lista

list(dados.Tipo.drop_duplicates())

### Criando uma Nova Lista
* pegamos a lista atual e descartamos os tipos que não queremos

residencial = ['Quitinete',
 'Casa',
 'Apartamento',
 'Casa de Condomínio',
 'Casa de Vila']

residencial

## Método isin()
* compara a variável e o dataframe passados
* cria uma series booleana, identificando se o que está na variável, consta no dataframe

dados.head(10)

dados.Tipo.isin(residencial).head(10)

* pode-se observar com os dois comandos acima que:
    * valores presentes na variável e no dataframe recebem True
    * valores não presentes recebem False

### Criando um Novo Dataframe

* colocando o valor acima em uma variável

selecao = dados.Tipo.isin(residencial)
selecao

* passando essa informação para um dataframe, ele vai manter apenas os registros que tiverem o valor True
* os valores com False são descartados

dados_residencial = dados[selecao]
dados_residencial

### Verificando se a Seleção foi Realizada Corretamente

list(dados_residencial.Tipo.drop_duplicates())

### Comparando o Tamanho dos Dataframes

print(dados_residencial.shape[0])
print(dados.shape[0])

## Reconstruindo o Índice

* atribuímos ao índice um range com o próprio tamanho dele

dados_residencial.index = range(dados_residencial.shape[0])
dados_residencial

# Exportando Dados
* muitas vezes queremos exportar um dataframe que criamos para utilizar em outros projetos
* para isso, assim como a função read_csv, temos a fução to_csv
* utilizamos também o separador ';'
* e marcamos o index como False, senão ele exporta o index e depois cria outro na importação

dados_residencial.to_csv('../dados/aluguel_residencial.csv', sep=';', index=False)

dados_residencial_2 = pd.read_csv('../dados/aluguel_residencial.csv', sep=';')
dados_residencial_2
