==> 02_fundamentos_pandas/notebook/10_selecoes_e_frequencia.ipynb <==
# Relatório de Análise IV

## Seleções e Frequência

Com base no último bando de dados criado, aluguel_residencial, criaremos o seguinte:

* Selecione somente os imóveis classificados com tipo 'Apartamento'.

* Selecione os imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'.

* Selecione os imóveis com área entre 60 e 100 metros quadrados, incluindo os limites.

* Selecione os imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00.

import pandas as pd

dados = pd.read_csv('../dados/aluguel_residencial.csv', sep=';')
dados.head(10)

## Selecione somente os imóveis classificados com tipo 'Apartamento'

# visualizando só apartamentos
selecao = dados['Tipo'] == 'Apartamento'
n1 = dados[selecao].head(10)
n1

# visualizando somente a frequência
n1 = dados[selecao].shape[0]
n1

## Selecione imóveis dos tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'

# quando há + de 1 termo usa-se o parênteses
# a barra '|' é o operador 'ou'
selecao = (dados['Tipo'] == 'Casa') | (dados['Tipo'] == 'Casa de Condomínio') | (dados['Tipo'] == 'Casa de Vila')
n2 = dados[selecao].shape[0]
n2

## Selecione imóveis com área entre 60 e 100 metros quadrados, incluindo os limites

# será usado parênteses novamente por ter mais de um termo
# nesse caso o operador '&' representa 'e'
selecao = (dados['Area'] >= 60) & (dados['Area'] <= 100)
n3 = dados[selecao].shape[0]
n3

## Selecione imóveis com pelo menos 4 quartos e aluguel menor que R$ 2.000,00

# nesse caso também será usado o operador 'e', '&'
# será usada duas variáveis (colunas) diferentes, 'Quartos' & 'Valor'
# o valor do aluguel deve ser menor de 2000
selecao = (dados['Quartos'] >= 4) & (dados['Valor'] < 2000) 
n4 = dados[selecao].shape[0]
n4

## Criando um print com os valores encontrados


## Seleções e Frequência

Com base no último bando de dados criado, aluguel_residencial, criaremos o seguinte:

* Selecione somente os imóveis classificados com tipo 'Apartamento'.

* Selecione os imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila'.

* Selecione os imóveis com área entre 60 e 100 metros quadrados, incluindo os limites.

* Selecione os imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00.print(f"Nº de imóveis classificados com tipo 'Apartamento' -> {n1}")

print(f"Nº de imóveis classificados com tipos 'Casa', 'Casa de Condomínio' e 'Casa de Vila' -> {n2}")

print(f"Nº de imóveis com área entre 60 e 100 metros quadrados, incluindo os limites -> {n3}")

print(f"Nº de imóveis que tenham pelo menos 4 quartos e aluguel menor que R$ 2.000,00 -> {n4}")

## Exercícios

# Considere o seguinte DataFrame para responder os próximos exercícios:

alunos = pd.DataFrame({'Nome': ['Ary', 'Cátia', 'Denis', 'Beto', 'Bruna', 'Dara', 'Carlos', 'Alice'], 
                        'Sexo': ['M', 'F', 'M', 'M', 'F', 'F', 'M', 'F'], 
                        'Idade': [15, 27, 56, 32, 42, 21, 19, 35], 
                        'Notas': [7.5, 2.5, 5.0, 10, 8.2, 7, 6, 5.6], 
                        'Aprovado': [True, False, False, True, True, True, False, False]}, 
                        columns = ['Nome', 'Idade', 'Sexo', 'Notas', 'Aprovado'])
alunos

### 1. Crie um DataFrame somente com os alunos aprovados

selecao = alunos['Aprovado'] == True
aprovados = alunos[selecao]
aprovados

### 2. Crie um DataFrame somente com as alunas aprovadas

selecao = (alunos['Sexo'] == 'F') & (alunos['Aprovado'] == True)
alunas_aprovadas = alunos[selecao]
alunas_aprovadas

### 3. Crie apenas uma visualização dos alunos com idade entre 10 e 20 anos ou com idade maior ou igual a 40 anos

# é pedido apenas para criar uma visualização
# portanto, não se deve colocar alunos[selecao] em uma outra variável, pois criará um outro dataframe
selecao = (alunos['Idade'] > 10) & (alunos['Idade'] < 20) | (alunos['Idade'] >= 40)
alunos[selecao]
