==> 02_fundamentos_pandas/notebook/13_tratamento_condicional.ipynb <==
# Relatório de Análise V

## Tratamento de dados faltantes

import pandas as pd

dados = pd.read_csv('../dados/aluguel_residencial.csv', sep=';')
dados.head(10)

## Método que verifica onde há valor nulo

dados.isnull()

## Método que verifica onde não há valor nulo

dados.notnull()

## Verificar Informações do Dataframe

# mostra a quantidade de registros que o dataframe possui
# mostra a quantidade de resgistros não nulos de cada variável
# podemos então identificar quais variáveis possuem registros nulos
dados.info()

* no caso acima, podemos ver que além das variáveis **Condomínio e IPTU**, a variável **Valor**, também possue alguns registros nulos

## Verificando apenas os valores nulos de **Valor**

# passando a seleção direto para o dataframe
dados[dados['Valor'].isnull()]

* a variável valor é a mais importante desse banco de dados
* portanto, não interessa manter esses registros nulos e serão removidos

## Removendo registros nulos

* para melhor visualização, serão criadas variáveis para verificar o tamanho do df antes e depois da remoção

A = dados.shape[0]
# recebe o tamanho do df antes da remoção
dados.dropna(subset = ['Valor'], inplace=True)
# remove valores nulos de uma variável ou lista de variáveis
# é necessário o inplace para execução no df
B = dados.shape[0]
# recebe o tamanho do df depois da remoção
A - B
# mostra a diferença para saber o tanto de registros removidos

# não retorna mais nenhum valor nulo
dados[dados.Valor.isnull()]

# Tratamento Condicional
## Tratamento Missings (continuação)

## Verificando quantos registros nulos temos em uma variável

dados[dados['Condominio'].isnull()].shape[0]

## Filtrando mais de uma variável

# cada filtro deve estar entre parênteses
selecao = (dados['Tipo'] == 'Apartamento') & (dados['Condominio'].isnull())

## Descartando a seleção que não interessa

A = dados.shape[0]
dados = dados[~selecao]
# esse comando com o '~' inverte a Series booleana
# então pegamos um valor que não interessa e descartamos
B = dados.shape[0]
A - B

# foram eliminados 9 registros que estavam com valor nulo
# mais 745 apartamentos com condomínio nulo
dados.shape[0]

# ainda permanecem registros nulos, mas não para apartamentos
dados[dados['Condominio'].isnull()].shape[0]

## Atribuindo valores para registros nulos

# substitui todos os registros nulos do dataframe por 0
# é necessário o inplace para gravar no dataframe
# fillna() é o método que preenche valores nulos
dados.fillna(0)

## Escolhendo variáveis e valores para registros nulos
### Usando um Dicionário
* com o dicionário é possível escolher quais variáveis serão preenchidas e quais valores serão atribuídos

dados = dados.fillna({'Condominio': 0, 'IPTU': 0})

### Verificando se há valores nulos

dados[dados['Condominio'].isnull()].shape[0]

dados[dados['Condominio'].isnull()].shape[0]

dados.info()

## Exportando e Sobrescrevendo o Dataframe

# para fins didáticos e manutenção dos notebooks, não sobrescreverei o arquivo
dados.to_csv('../dados/aluguel_residencial_notnull.csv', sep=';', index = False)
