==> 02_fundamentos_pandas/notebook/14_metodos_de_interpolacao.ipynb <==
# Métodos de Interpolação

import pandas as pd

## Criando um banco qualquer

data = [0.5, None, None, 0.52, 0.54, None, None, 0.59, 0.6, None, 0.7]

## Criando uma Series com esses dados

s = pd.Series(data)
s

## Prrenchendo os valores nulos com zero

# pode-se adicionar qualquer valor dentro dos parênteses, até mesmo strings
s.fillna(0)

## Outros Métodos de Interpolação

# observando a series original
s

### Método ffill
* analisa a Series do primeiro elemento ao último
* preenche o registro nulo com o valor anteriormente válido
* método que pode ser usado com dados de uma série temporal
* quando há um valor nulo, as vezes é razoável preencher com o do ano anterior por exemplo
* isso permite que as funções sejam executadas sem problemas

# esse método preenche o registro nulo com o valor anteriormente válido
s.fillna(method = 'ffill')

### Método bfill
* funciona exatamente como o método ffill, mas ao contrário
* percorre do último ao primeiro elemento
* preenche entradas nulas com o valor válido anterior
* nesse caso o valor anterior é o valor de baixo

s.fillna(method = 'bfill')

### Preenchendo com a média de valores válidos

# observando a series original
s

s.fillna(s.mean())

## Observações

* algumas vezes esses métodos de interpolação podem gerar gráficos discrepantes
* principalmente gráficos representados por curvas, podem gerar retas, onde há a interpolação
* por isso pode-se limitar o nº de preenchimentos com o parâmetro 'limit'
* também é possível mesclar os métodos para suavizar esse impacto

## Limitando a interpolação

# observando a series original
s

# limitando a um preenchimento
s.fillna(method = 'ffill', limit = 1)

## Mesclando os Métodos
* mesclando ffill e bfill
* dessa forma suavizamos os dados, preenchendo 1º com os dados acima e depois com os dados válidos de baixo

s1 = s.fillna(method = 'ffill', limit = 1)
s1

s1.fillna(method = 'bfill', limit = 1)
