==> 02_fundamentos_pandas/notebook/15_criando_novas_variaveis.ipynb <==
# Relatório de Análise VI

## Criando Novas Variáveis

import pandas as pd

dados = pd.read_csv('../dados/aluguel_residencial_notnull.csv', sep=';')
dados.head(10)

## Criando uma Variável (coluna)

### Somando valores de outras colunas

dados['Valor Bruto'] = dados['Valor'] + dados['Condominio'] + dados['IPTU']
dados.head(10)

### Usando Divisão

dados['Valor m2'] = dados['Valor'] / dados['Area']
dados.head(10).sort_index(axis=1, ascending=False)
# para melhor visualização o cabeçalho da coluna foi organizado de trás para frente
# para não cortar a coluna quando exportar o arquivo para pdf
# utilizou-se então a função sort_index, com os parâmetros:
# axis=1 => para organizar as colunas
# ascending=False => para organizar de maneira decrescente

### Arrendondando as casas decimais

dados['Valor m2'] = dados['Valor m2'].round(2)
dados.head(10).sort_index(axis=1, ascending=False)
# também é possível arredondar diretamente, colocando tudo entre parênteses
# dados['Valor m2'] = (dados['Valor'] / dados['Area']).round(2)

### Continuando com a divisão

dados['Valor Bruto m2'] = (dados['Valor Bruto'] / dados['Area']).round(2)
dados.head(10).sort_index(axis=1, ascending=False)

## Filtrando Valores para Criar uma Nova Variável

### Listando quais valores existem em Tipo

list(dados['Tipo'].unique())

### Selecionando a variável desejada

casa = ['Casa', 'Casa de Condomínio', 'Casa de Vila']

### Criando a nova coluna com aplicação de uma função lambda para cada registro

dados['Tipo Agregado'] = dados['Tipo'].apply(lambda x: 'Casa' 
                                             if x in casa else 'Apartamento')
dados.head(10).sort_index(axis=1, ascending=False)

# Verificando se Casa de Condomínio também recebeu o tratamento certo
dados[dados['Tipo'] == 'Casa de Condomínio'].sort_index(axis=1, 
                                                        ascending=False).head(10)

### Aplicando a função lambda com várias condicionais

# cuidado com os parênteses
# colocar as outras condicionais dentro de parênteses do primeiro else
# e cada if também entre parênteses
dados['Tamanho'] = dados['Quartos'].apply(lambda x: 'Grande' if x > 3 
                                          else ('Médio' if (x >= 2 and x <= 3) 
                                                else 'Pequeno'))
dados.head(10).sort_index(axis=1, ascending=False)

# exportando um novo arquivo para trabalhar no próximo notebook
dados.to_csv('../dados/aluguel_res_novas_variaveis.csv', sep=';', index=False)
