==> 02_fundamentos_pandas/notebook/16_excluindo_variaveis.ipynb <==
# Excluindo Variáveis do Dataframe

import pandas as pd

dados = pd.read_csv('../dados/aluguel_res_novas_variaveis.csv', sep=';')

dados.head(10)

dados_aux = dados[['Tipo Agregado', 'Valor Bruto', 'Valor Bruto m2', 'Valor m2', 'Tamanho']]
dados_aux.head(10)

## Formas de excluir variáveis

# usando del
del dados_aux['Tipo Agregado']
dados_aux.head(10)

# usando pop
dados_aux.pop('Valor Bruto')
dados_aux.head(10)

# verificando o nome das colunas
dados.columns

# Também é possível colacar em uma lista
list(dados.columns)

# usando drop, é possível passar várias variáveis
# também é possível excluir linhas
# deve-se passar o eixo e o inplace para executar no df
dados.drop(['Valor Bruto', 'Valor Bruto m2', 'Tamanho'], axis=1, inplace=True)
dados.head(10)

# exportando novamente para utilizar nos próximos notebooks
dados.to_csv('../dados/aluguel_res_del_variaveis.csv', sep=';', index=False)
