==> 02_fundamentos_pandas/notebook/18_criando_agrupamentos.ipynb <==
# Criando Agrupamentos

import pandas as pd

dados = pd.read_csv('../dados/aluguel_res_del_variaveis.csv', sep=';')

dados.head(10)

# mostra a média de todos os valores da coluna Valor
# arredondando para 2 casas decimais
round(dados.Valor.mean(), 2)

### Consultar documentação de Dataframe do pandas
#### https://pandas.pydata.org/pandas-docs/stable/reference/frame.html

## Verificar os bairros do dataframe

dados['Bairro'].unique()

## Fazendo uma seleção com o método isin()

bairros = ['Barra da Tijuca', 'Copacabana', 'Ipanema', 'Leblon', 'Botafogo', 'Flamengo', 'Tijuca']
selecao = dados['Bairro'].isin(bairros)
dados = dados[selecao]

dados['Bairro'].drop_duplicates()

#### Também é possível fazer dessa forma, mas o comando fica enorme dependendo da quantidade de valores
##### Repare que mesmo usando apenas 2 valores de exemplo é fácil se perder com o comando

selecao2 = (dados['Bairro'] == 'Copacabana') | (dados['Bairro'] == 'Tijuca')
dados2 = dados[selecao2]
dados2['Bairro'].drop_duplicates()

## Agrupando com groupby

* para usar o groupby é necessário passar a variável desejada dentro dos parênteses
* é criado um objeto do tipo DataFrameGroupBy, do pandas
* esse método cria um dicionário, com uma indexação baseada no valor da variável e o index que cada valor foi encontrado no dataframe

grupo_bairro = dados.groupby('Bairro')
type(grupo_bairro)

# a chave é o valor (nome) da variável
# a lista traz o index de cada valor encontrado no dataframe
grupo_bairro.groups

## Criando uma média para cada bairro

# essa variável dados refere-se só ao laço
# ela guarda os valores do dataframe de cada bairro
for bairro, dados in grupo_bairro:
    print(f'{bairro} -> {dados.Valor.mean():.2f}')

## Tirando a média usando o groupby

# fica muito mais simples usando o agrupamento
grupo_bairro.Valor.mean().round(2)

# também é possível passar mais de uma variável
grupo_bairro[['Valor', 'Condominio']].mean().round(2)

## Exercício

Considere o seguinte DataFrame para responder o exercício abaixo:

import pandas as pd
alunos = pd.DataFrame({'Nome': ['Ary', 'Cátia', 'Denis', 'Beto', 'Bruna', 'Dara', 'Carlos', 'Alice'], 
                        'Sexo': ['M', 'F', 'M', 'M', 'F', 'F', 'M', 'F'], 
                        'Idade': [15, 27, 56, 32, 42, 21, 19, 35], 
                        'Notas': [7.5, 2.5, 5.0, 10, 8.2, 7, 6, 5.6], 
                        'Aprovado': [True, False, False, True, True, True, False, False]}, 
                        columns = ['Nome', 'Idade', 'Sexo', 'Notas', 'Aprovado'])

alunos

Como devemos proceder para obter um **DataFrame** com as notas médias dos alunos, com duas casas decimais, segundo seu sexo?

sexo = alunos.groupby('Sexo')
sexo = pd.DataFrame(sexo['Notas'].mean().round(2))
sexo.columns = ['Notas Médias']
sexo
