==> 02_fundamentos_pandas/notebook/20_criando_faixas_de_valor.ipynb <==
# Criando Faixas de Valor

import pandas as pd

dados = pd.read_csv('../dados/aluguel2.csv', sep=';')
dados.head(10)

## Distribuição de Frequência

### Faixas de Frequência de Quartos

* 1 e 2
* 3 e 4
* 5 e 6
* 7 ou mais

# verificando o número máximo de quartos que aparece
dados['Quartos'].max()

### Criando Classes

# criamos as classes com o valor mínimo, máximo e os limites
classes = [0, 2, 4, 6, 14]

#### Categorizando as classes

# utiliza-se 2 argumentos
# nesse caso está sendo utilizado uma series e uma lista com os limites
quartos = pd.cut(dados.Quartos, classes)
quartos

# o parênteses não inclui o valor
# o colchetes inclui o valor
# ficando as faixas de quartos conforme estabelecido acima
pd.value_counts(quartos)

#### Modificando a visualização

labels = ['1 e 2 quartos', '3 e 4 quartos', '5 e 6 quartos', '7 quartos ou mais']

# pode-se adicionar mais um parâmetro
# nesse caso utilizamos o labels
quartos = pd.cut(dados.Quartos, classes, labels=labels)

pd.value_counts(quartos)

* sabendo que é possível que não tenha sido informado o número de quartos, essa informação pode ser 0
* podemos incluir então um quarto parâmetro

# include_lowest inclui o menor valor possível que aparece
quartos = pd.cut(dados.Quartos, classes, labels=labels, include_lowest=True)

# pode-se notar que há uma diferença na faixa 1 e 2 quartos
# ou seja, foi incluido o valor 0 também
pd.value_counts(quartos)
